# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for BloodConnect donor, blood request, donation, and hospital services.",
        "title": "BloodConnect Donor Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1324",
    "basePath": "/api/v1",
    "paths": {
...
        "/donor-matches/{id}/route": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get route information for a donor match using Geoapify",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Geoapify"
                ],
                "summary": "Get route for donor match",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Donor Match ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.GeoapifyRouteSwaggoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    }
                }
            }
        },
        "/hospitals": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of hospitals from Geoapify based on city",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Geoapify"
                ],
                "summary": "Get hospitals by city",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Bekasi",
                        "description": "City name",
                        "name": "city",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.GeoapifyHospitalListSwaggoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.GeoapifyHospital": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "domain.GeoapifyRoute": {
            "type": "object",
            "properties": {
                "distance": {
                    "type": "number"
                },
                "distance_units": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "time": {
                    "type": "number"
                }
            }
        },
...
        "helper.ErrorDetailSwaggoResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "helper.ErrorSwaggoResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/helper.ErrorDetailSwaggoResponse"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "helper.GeoapifyHospitalListSwaggoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.GeoapifyHospital"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "helper.GeoapifyRouteSwaggoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.GeoapifyRoute"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}